finish the course

+++ node-todo-api/server/server.js
@@ -54,21 +54,39 @@ app.get('/todos/:id', authenticate, (request, response)=> {
     
 });
 
-app.delete('/todos/:id', authenticate, (request, response)=> {
-    let id = request.params.id;
-    if (!ObjectID.isValid(id)) {
-        return response.status(404).send('Id not valid');
-    }
-    Todo.findOneAndDelete({
-        _id: id,
-        _creator: request.user._id
-    }).then((todo)=> {
+// app.delete('/todos/:id', authenticate, (request, response)=> {
+//     let id = request.params.id;
+//     if (!ObjectID.isValid(id)) {
+//         return response.status(404).send('Id not valid');
+//     }
+//     Todo.findOneAndDelete({
+//         _id: id,
+//         _creator: request.user._id
+//     }).then((todo)=> {
+//         if (!todo) {
+//             return response.status(404).send('No todo found');
+//         }
+//         response.send({todo});
+//     }).catch(error=> response.status(400).send());
+// });
+
+app.delete('/todos/:id', authenticate, async (request, response)=> {
+    try {
+        const id = request.params.id;
+        if (!ObjectID.isValid(id)) {
+            return response.status(404).send('Id not valid');
+        }
+        const todo = await Todo.findOneAndDelete({_id: id, _creator: request.user._id});
         if (!todo) {
             return response.status(404).send('No todo found');
         }
         response.send({todo});
-    }).catch(error=> response.status(400).send());
+    } catch(error) {
+        response.status(400).send();
+    }
 });
+
+
 
 app.patch('/todos/:id', authenticate, (request, response)=> {
     let id = request.params.id;
@@ -97,16 +115,29 @@ app.patch('/todos/:id', authenticate, (request, response)=> {
     });
 });
 
-app.post('/users', (request, response)=> {
-    let body = _.pick(request.body, ['email', 'password']);
-    let user = new User(body);
-    user.save().then(()=> {
-        return user.generateAuthToken();
-    }).then((token)=> {
+// app.post('/users', (request, response)=> {
+//     let body = _.pick(request.body, ['email', 'password']);
+//     let user = new User(body);
+//     user.save().then(()=> {
+//         return user.generateAuthToken();
+//     }).then((token)=> {
+//         response.header('x-auth', token).send(user);
+//     }).catch(error=> {
+//         response.status(400).send(error);
+//     });
+// });
+
+app.post('/users', async (request, response)=> {
+    try {
+        const body = _.pick(request.body, ['email', 'password']);
+        const user = new User(body);
+        await user.save();
+        const token = await user.generateAuthToken();
         response.header('x-auth', token).send(user);
-    }).catch(error=> {
+    } catch(error) {
         response.status(400).send(error);
-    });
+    }
+    
 });
 
 
@@ -115,24 +146,42 @@ app.get('/users/me', authenticate, (request, response)=> {
     response.send(request.user);
 });
 
-app.post('/users/login', (request, response)=> {
-    let body = _.pick(request.body, ['email', 'password']);
-    User.findByCredentials(body.email, body.password).then((user)=> {
+// app.post('/users/login', (request, response)=> {
+//     let body = _.pick(request.body, ['email', 'password']);
+//     User.findByCredentials(body.email, body.password).then((user)=> {
         
-        return user.generateAuthToken().then((token)=> {
-            response.header('x-auth', token).send(user);
-        });
-    }).catch((error)=> {
+//         return user.generateAuthToken().then((token)=> {
+//             response.header('x-auth', token).send(user);
+//         });
+//     }).catch((error)=> {
+//         response.status(400).send();
+//     });    
+// });
+app.post('/users/login', async (request, response)=> {
+    try {
+        const body = _.pick(request.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken();
+        response.header('x-auth', token).send(user);
+    } catch(e) {
         response.status(400).send();
-    });    
+    }
 });
 
-app.delete('/users/me/token', authenticate, (request, response)=> {
-    request.user.removeToken(request.token).then(()=> {
+// app.delete('/users/me/token', authenticate, (request, response)=> {
+//     request.user.removeToken(request.token).then(()=> {
+//         response.status(200).send();
+//     }).catch((error)=> {
+//         response.status(400).send();
+//     });
+// });
+app.delete('/users/me/token', authenticate, async (request, response)=> {
+    try {
+        await request.user.removeToken(request.token);
         response.status(200).send();
-    }).catch((error)=> {
+    } catch(e) {
         response.status(400).send();
-    });
+    }
 });
 
 app.listen(process.env.PORT, process.env.IP, ()=> {
